fixed bugs

+++ src/Codegen/CodeGenPass.cpp
@@ -875,7 +875,7 @@ llvm::Value *CodeGenPass::visitTupleLiteral(TupleLiteral *TupleLit) {
     for (auto Child : *TupleLit) {
         auto MemberVal = visit(Child);
         auto MemLoc = IR.CreateGEP(
-                TupLoc, {IR.getInt64(0), IR.getInt64(CurrIdx++)});
+                TupLoc, {IR.getInt32(0), IR.getInt32(CurrIdx++)});
         IR.CreateStore(MemberVal, MemLoc);
     }
     return IR.CreateLoad(TupLoc);
@@ -1364,7 +1364,7 @@ llvm::Value *CodeGenPass::visitMemberReference(MemberReference *Ref) {
         throw std::runtime_error("Only int literals should reach here");
     auto StructLoc = SymbolMap[Ref->getIdentifier()->getReferred()];
     return IR.CreateGEP(StructLoc, {
-        IR.getInt64(0), IR.getInt64(MemIdx->getVal() - 1)});
+        IR.getInt32(0), IR.getInt32(MemIdx->getVal() - 1)});
 }
 
 llvm::Function *CodeGenPass::getOrInsertFunction(Type *Ty,
